+++ soccer_common/src/soccer_common/__init__.py
@@ -2,7 +2,7 @@ from soccer_common.transformation import Transformation
 from soccer_common.transformation2d import Transformation2D
 
 try:
-    from soccer_common.camera import Camera
+    from soccer_common.perception.camera import Camera
     from soccer_common.pid import PID
 except:
     print("not using camera")
