+++ soccer_strategy/src/soccer_strategy/old/robot_controlled_2d.py
@@ -5,7 +5,7 @@ import numpy as np
 import rospy
 from soccer_pycontrol import path
 
-from soccer_common.camera import Camera
+from soccer_common.perception.camera import Camera
 from soccer_common.transformation import Transformation
 from soccer_common.utils import wrapToPi
 from soccer_strategy.old.ball import Ball
